피해량 및 치유량 관련 나눗셈 연산이 double 값을 반환하도록 수정

+++ src/main/java/com/dace/dmgr/combat/character/neace/Neace.java
@@ -156,7 +156,7 @@ public final class Neace extends Support {
         super.onGiveHeal(provider, target, amount);
 
         if (provider != target && target instanceof CombatUser)
-            provider.addScore("치유", (double) (HEAL_SCORE * amount) / target.getDamageModule().getMaxHealth());
+            provider.addScore("치유", HEAL_SCORE * amount / target.getDamageModule().getMaxHealth());
 
         return true;
     }
